HoangNh: them cau hinh

+++ lib/presentation/dashboard_screen/page/student_management/widget/user_sources.dart
@@ -218,13 +218,33 @@ Widget buildActive(
       child: CusstomActiveTable(
         showView: true,
         onDelete: () {
-          controller.deleterUser(ma_sinh_vien);
-          Get.dialog(Dialog(
-            child: CustomLoading(),
-          ));
-          Future.delayed(Duration(seconds: 2),() {
-            Get.back();
-          },);
+           Get.dialog(
+            AlertDialog(
+              content: Container(
+                height: 100,
+                width: 100,
+                alignment: Alignment.center,
+                child: Text('Bạn có chắc muốn xoá không?')
+                ),
+              actions: [
+                CustomButtonAlert(
+                  titileDisable: 'Huỷ',
+                  titileEnable: 'Đồng ý',
+                  onPressedDisable: () => Get.back(),
+                  onPressedEnable: () {
+                    Get.back();
+                    controller.deleterUser(ma_sinh_vien);
+                    Get.dialog(Dialog(
+                      child: CustomLoading(),
+                    ));
+                    Future.delayed(Duration(seconds: 2),() {
+                      Get.back();
+                    },);
+                  },
+                )
+              ],
+            )
+          );
         },
         onUpdate: () {
           Get.dialog(alertAvt(
